requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
pydantic
requests
streamlit
beautifulsoup4
